refactor: migrate normalization and configuration functions to openaiApi.js
- Moved normalization functions for base URL and model name, along with the configuration retrieval logic, from openaiCompatibleTranslate.js to a new utility file openaiApi.js.
- Removed redundant functions from openaiCompatibleTranslate.js to streamline the codebase.
- Updated logging to ensure consistent behavior across the new utility functions.
