+++ translation_providers/openaiCompatibleTranslate.js
@@ -1,7 +1,7 @@
 // disneyplus-dualsub-chrome-extension/translation_providers/openaiCompatibleTranslate.js
 
 import Logger from '../utils/logger.js';
-import { configService } from '../services/configService.js';
+import { getConfig, normalizeBaseUrl, normalizeModelName } from '../utils/openaiApi.js';
 
 // Initialize logger for OpenAI-compatible translation provider
 const logger = Logger.create('OpenAICompatibleTranslate');
@@ -15,55 +15,6 @@ const TOKEN_MULTIPLIER = 3; // Approximate tokens per character ratio for text e
 const MAX_TOKENS = 4000;
 const ERROR_TEXT_PREVIEW_LENGTH = 500;
 const ERROR_TEXT_TRUNCATION_LIMIT = 200;
-
-/**
- * Normalizes baseUrl by removing trailing slashes and backslashes
- * @param {string} url - The base URL to normalize
- * @returns {string} Normalized URL without trailing slashes
- */
-function normalizeBaseUrl(url) {
-    if (!url || typeof url !== 'string') {
-        return url;
-    }
-    
-    // Remove trailing slashes and backslashes
-    const normalized = url.replace(/[/\\]+$/, '');
-    
-    logger.debug('Base URL normalized', {
-        originalUrl: url,
-        normalizedUrl: normalized,
-        hadTrailingSlash: url !== normalized
-    });
-    
-    return normalized;
-}
-
-/**
- * Normalizes model name for OpenAI-compatible endpoints
- * @param {string} model - The model name to normalize
- * @param {string} baseUrl - The base URL to determine provider type
- * @returns {string} Normalized model name
- */
-function normalizeModelName(model, baseUrl) {
-    if (!model || typeof model !== 'string') {
-        return model;
-    }
-    
-    // For Google Gemini API, remove "models/" prefix if present
-    if (baseUrl && baseUrl.includes('generativelanguage.googleapis.com')) {
-        const normalized = model.startsWith('models/') ? model.substring(7) : model;
-        
-        logger.debug('Model name normalized for Gemini', {
-            originalModel: model,
-            normalizedModel: normalized,
-            hadModelsPrefix: model.startsWith('models/')
-        });
-        
-        return normalized;
-    }
-    
-    return model;
-}
 
 /**
  * Translates text using OpenAI-compatible API with Gemini models.
@@ -331,274 +282,5 @@ export async function translate(text, sourceLang, targetLang) {
         
         // Re-throw the error to be caught by the caller
         throw error;
-    }
-}
-
-/**
- * Retrieves the configuration from storage using the config service.
- * @returns {Promise<Object>} Configuration object with apiKey, baseUrl, and model.
- */
-async function getConfig() {
-    logger.debug('Retrieving configuration via configService');
-    
-    const config = await configService.getMultiple([
-        'openaiCompatibleApiKey',
-        'openaiCompatibleBaseUrl', 
-        'openaiCompatibleModel'
-    ]);
-
-    const result = {
-        apiKey: config.openaiCompatibleApiKey,
-        baseUrl: config.openaiCompatibleBaseUrl,
-        model: config.openaiCompatibleModel
-    };
-
-    logger.debug('Configuration retrieved successfully via configService', {
-        hasApiKey: !!result.apiKey,
-        baseUrl: result.baseUrl,
-        model: result.model
-    });
-
-    return result;
-}
-
-/**
- * Sets the API key using the config service.
- * @param {string} apiKey The API key to store.
- * @returns {Promise<void>}
- */
-export async function setApiKey(apiKey) {
-    logger.debug('Setting API key via configService');
-    await configService.set('openaiCompatibleApiKey', apiKey);
-    logger.info('API key stored successfully');
-}
-
-/**
- * Sets the base URL using the config service.
- * @param {string} baseUrl The base URL to store.
- * @returns {Promise<void>}
- */
-export async function setBaseUrl(baseUrl) {
-    // Normalize the URL before storing
-    const normalizedUrl = normalizeBaseUrl(baseUrl);
-    
-    logger.debug('Setting base URL via configService', {
-        originalUrl: baseUrl,
-        normalizedUrl: normalizedUrl
-    });
-    
-    await configService.set('openaiCompatibleBaseUrl', normalizedUrl);
-    logger.info('Base URL stored successfully');
-}
-
-/**
- * Sets the model using the config service.
- * @param {string} model The model to store.
- * @returns {Promise<void>}
- */
-export async function setModel(model) {
-    logger.debug('Setting model via configService', { model });
-    await configService.set('openaiCompatibleModel', model);
-    logger.info('Model stored successfully', { model });
-}
-
-/**
- * Fetches available models from the API.
- * @returns {Promise<Array>} Array of available models.
- */
-export async function getAvailableModels() {
-    try {
-        logger.debug('Fetching available models from API');
-        
-        const config = await getConfig();
-        if (!config.apiKey) {
-            const error = new Error('API key not configured');
-            logger.error('Cannot fetch models: API key not configured', error);
-            throw error;
-        }
-
-        const normalizedBaseUrl = normalizeBaseUrl(config.baseUrl) || 'https://generativelanguage.googleapis.com/v1beta/openai';
-        const modelsUrl = `${normalizedBaseUrl}/models`;
-
-        logger.debug('Making models API request', {
-            modelsUrl: modelsUrl,
-            hasApiKey: !!config.apiKey
-        });
-
-        const response = await fetch(modelsUrl, {
-            method: 'GET',
-            headers: {
-                'Authorization': `Bearer ${config.apiKey}`
-            }
-        });
-
-        if (!response.ok) {
-            const errorText = await response.text();
-            logger.error('Models API HTTP error', null, {
-                status: response.status,
-                statusText: response.statusText,
-                responsePreview: errorText.substring(0, 200),
-                modelsUrl: modelsUrl
-            });
-            throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-        }
-
-        const data = await response.json();
-        const models = data.data || [];
-        
-        logger.info('Available models fetched successfully', {
-            modelCount: models.length,
-            modelNames: models.map(m => m.id || m.name).slice(0, 5) // Log first 5 model names
-        });
-        
-        return models;
-    } catch (error) {
-        logger.error('Error fetching available models', error);
-        throw error;
-    }
-}
-
-/**
- * Tests the API connection and key validity using lightweight endpoints.
- * @returns {Promise<boolean>} True if the API key is valid and working.
- */
-export async function testConnection() {
-    try {
-        logger.debug('Testing API connection');
-        
-        const config = await getConfig();
-        if (!config.apiKey) {
-            logger.warn('Cannot test connection: API key not configured');
-            return false;
-        }
-
-        const normalizedBaseUrl = normalizeBaseUrl(config.baseUrl) || 'https://generativelanguage.googleapis.com/v1beta/openai';
-        
-        // Method 1: Try /models endpoint (lightweight and doesn't consume quota)
-        try {
-            const models = await getAvailableModels();
-            if (models && models.length > 0) {
-                logger.info('Connection test successful via models API', {
-                    modelCount: models.length
-                });
-                return true;
-            }
-        } catch (modelsError) {
-            logger.debug('Models endpoint failed, trying alternative methods', {
-                error: modelsError.message
-            });
-        }
-        
-        // Method 2: Try HEAD request to base URL (most lightweight)
-        try {
-            logger.debug('Testing connection with HEAD request', {
-                baseUrl: normalizedBaseUrl
-            });
-            
-            const headResponse = await fetch(normalizedBaseUrl, {
-                method: 'HEAD',
-                headers: {
-                    'Authorization': `Bearer ${config.apiKey}`
-                }
-            });
-            
-            // Consider 2xx, 4xx responses as connection success (server is reachable)
-            // 401/403 means server is reachable but auth issue
-            if (headResponse.status < 500) {
-                logger.info('Connection test successful via HEAD request', {
-                    status: headResponse.status,
-                    statusText: headResponse.statusText
-                });
-                return true;
-            }
-        } catch (headError) {
-            logger.debug('HEAD request failed, trying health endpoint', {
-                error: headError.message
-            });
-        }
-        
-        // Method 3: Try common health endpoints
-        const healthEndpoints = ['/health', '/v1/health', '/status'];
-        
-        for (const healthPath of healthEndpoints) {
-            try {
-                const healthUrl = `${normalizedBaseUrl}${healthPath}`;
-                logger.debug('Testing health endpoint', { healthUrl });
-                
-                const healthResponse = await fetch(healthUrl, {
-                    method: 'GET',
-                    headers: {
-                        'Authorization': `Bearer ${config.apiKey}`
-                    }
-                });
-                
-                if (healthResponse.ok) {
-                    logger.info('Connection test successful via health endpoint', {
-                        healthUrl: healthUrl,
-                        status: healthResponse.status
-                    });
-                    return true;
-                }
-            } catch (healthError) {
-                logger.debug('Health endpoint failed', {
-                    endpoint: healthPath,
-                    error: healthError.message
-                });
-            }
-        }
-        
-        // Method 4: Minimal API test (avoid translation to save quota)
-        // Try a very simple chat completion with minimal tokens
-        try {
-            logger.debug('Testing with minimal API call');
-            
-            const model = normalizeModelName(config.model || 'gemini-1.5-flash', normalizedBaseUrl);
-            const testUrl = `${normalizedBaseUrl}/chat/completions`;
-            
-            const minimalRequest = {
-                model: model,
-                messages: [
-                    {
-                        role: "user",
-                        content: "Hi"
-                    }
-                ],
-                max_tokens: 1, // Minimal token usage
-                temperature: 0
-            };
-            
-            const testResponse = await fetch(testUrl, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${config.apiKey}`
-                },
-                body: JSON.stringify(minimalRequest)
-            });
-            
-            if (testResponse.ok) {
-                logger.info('Connection test successful via minimal API call', {
-                    status: testResponse.status
-                });
-                return true;
-            } else {
-                logger.debug('Minimal API call failed', {
-                    status: testResponse.status,
-                    statusText: testResponse.statusText
-                });
-            }
-        } catch (apiError) {
-            logger.debug('Minimal API call failed', {
-                error: apiError.message
-            });
-        }
-        
-        // If all methods fail, log warning and return false
-        logger.warn('All connection test methods failed - API may be unreachable or misconfigured');
-        return false;
-        
-    } catch (error) {
-        logger.error('Connection test failed with unexpected error', error);
-        return false;
     }
 }
